scripts/inference.py
import yaml
import argparse

from unsloth import FastLanguageModel

from utils.logger import get_logger
from utils.paths import BASE_DIR, CONFIG_DIR

logger = get_logger(__name__)

class IntentClassification:
    def __init__(self, configs):
        self.prompt = """Here is a banking intent:
        {}
        
        Classify this banking intent into one label:
        01 to 77
        
        SOLUTION
        """
        
        self.label_map = {
            "1": "activate_my_card",
            "2": "age_limit",
            "3": "apple_pay_or_google_pay",
            "4": "atm_support",
            "5": "automatic_top_up",
            "6": "balance_not_updated_after_bank_transfer",
            "7": "balance_not_updated_after_cheque_or_cash_deposit",
            "8": "beneficiary_not_allowed",
            "9": "cancel_transfer",
            "10": "card_about_to_expire",
            "11": "card_acceptance",
            "12": "card_arrival",
            "13": "card_delivery_estimate",
            "14": "card_linking",
            "15": "card_not_working",
            "16": "card_payment_fee_charged",
            "17": "card_payment_not_recognised",
            "18": "card_payment_wrong_exchange_rate",
            "19": "card_swallowed",
            "20": "cash_withdrawal_charge",
            "21": "cash_withdrawal_not_recognised",
            "22": "change_pin",
            "23": "compromised_card",
            "24": "contactless_not_working",
            "25": "country_support",
            "26": "declined_card_payment",
            "27": "declined_cash_withdrawal",
            "28": "declined_transfer",
            "29": "direct_debit_payment_not_recognised",
            "30": "disposable_card_limits",
            "31": "edit_personal_details",
            "32": "exchange_charge",
            "33": "exchange_rate",
            "34": "exchange_via_app",
            "35": "extra_charge_on_statement",
            "36": "failed_transfer",
            "37": "fiat_currency_support",
            "38": "get_disposable_virtual_card",
            "39": "get_physical_card",
            "40": "getting_spare_card",
            "41": "getting_virtual_card",
            "42": "lost_or_stolen_card",
            "43": "lost_or_stolen_phone",
            "44": "order_physical_card",
            "45": "passcode_forgotten",
            "46": "pending_card_payment",
            "47": "pending_cash_withdrawal",
            "48": "pending_top_up",
            "49": "pending_transfer",
            "50": "pin_blocked",
            "51": "receiving_money",
            "52": "Refund_not_showing_up",
            "53": "request_refund",
            "54": "reverted_card_payment?",
            "55": "supported_cards_and_currencies",
            "56": "terminate_account",
            "57": "top_up_by_bank_transfer_charge",
            "58": "top_up_by_card_charge",
            "59": "top_up_by_cash_or_cheque",
            "60": "top_up_failed",
            "61": "top_up_limits",
            "62": "top_up_reverted",
            "63": "topping_up_by_card",
            "64": "transaction_charged_twice",
            "65": "transfer_fee_charged",
            "66": "transfer_into_account",
            "67": "transfer_not_received_by_recipient",
            "68": "transfer_timing",
            "69": "unable_to_verify_identity",
            "70": "verify_my_identity",
            "71": "verify_source_of_funds",
            "72": "verify_top_up",
            "73": "virtual_card_not_working",
            "74": "visa_or_mastercard",
            "75": "why_verify_identity",
            "76": "wrong_amount_of_cash_received",
            "77": "wrong_exchange_rate_for_cash_withdrawal"
        }
        
        self.configs = configs
        
        self.model, self.tokenizer = FastLanguageModel(**self.configs["model"])
        
    def __call__(self, message):
        cleaned_message = " ".join(message.lower().split())
        prompt = self.prompt.format(cleaned_message)
        encode = self.tokenizer(
            prompt,
            **self.configs["tokenizer"]
        )
        output = self.model.generate(
            **encode, 
            **self.configs["generate"]
        )
        decode = self.tokenizer.decode(output[:, -2:])[0]
        print(self.label_map[decode])
        
def main():
    try:
        parser = argparse.ArgumentParser()
        parser.add_argument("--message")
        
        args = parser.parse_args()
        message = args.message
        print(message)
        
        with open(CONFIG_DIR / "inference.yaml", "r") as f:
            configs = yaml.safe_load(f)
        
        configs["model"]["model_name"] = BASE_DIR / "models" / "unsloth" "Qwen3-4B-Base" / "finetuned"
        
        classifier = IntentClassification(configs)
        classifier(message)
    except Exception as e:
        logger.error(f"Inference process failed: {e}")
    
if __name__ == "__main__":
    main()