requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
bitsandbytes==0.49.2
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==4.3.0
debugpy==1.8.20
decorator==5.2.1
diffusers==0.37.1
dill==0.4.0
docstring_parser==0.18.0
dotenv==0.9.9
executing==2.2.1
filelock==3.29.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.4.3
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
importlib_metadata==9.0.0
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.21.1
multidict==6.7.1
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
packaging==26.1
pandas==3.0.2
parso==0.8.6
peft==0.19.1
pillow==12.2.0
platformdirs==4.9.6
polars==1.40.1
polars-runtime-32==1.40.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pure_eval==0.2.3
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchao==0.17.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.0
triton-windows==3.6.0.post26
trl==0.24.0
typeguard==4.5.1
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.13
tzdata==2026.1
unsloth==2026.4.7
unsloth_zoo==2026.4.9
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.47.0
xformers==0.0.35
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
